session-1/database/implementation/requirements.txt
mysql-connector-python==8.0.28
python-dotenv>=0.20.0

